backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-dotenv>=1.0.0
websockets>=12.0
pydantic>=2.5.0
aiofiles>=23.2.1
python-multipart>=0.0.6
httpx>=0.26.0
google-generativeai>=0.8.0
openai>=1.12.0
playwright>=1.41.0
beautifulsoup4>=4.12.0
markdown>=3.5.0
